requirements.txt
django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-redis==5.4.0
confluent-kafka==2.6.1
pymongo==4.10.1
# cassandra-driver — install in Phase 3 (not supported on Python 3.13 yet)
drf-spectacular==0.27.2
python-dotenv==1.0.1
PyJWT==2.9.0
django-cors-headers==4.9.0
pytest==8.3.4
pytest-django==4.9.0
